Move sig verification stage into its own module
